refactor: group admin form builder components into subdirectories
a little cleaner
