--- frontend/src/features/admin-form-builder/CreateFieldDrawer/index.ts
@@ -1,1 +1,0 @@
-export { CreateFieldDrawer } from './CreateFieldDrawer'
