+++ frontend/src/features/admin-form-builder/components/BuilderContent/index.ts
@@ -1,0 +1,1 @@
+export * from './BuilderContent'
