+++ frontend/src/features/admin-form-builder/components/BuilderDrawer/BuilderDrawer.tsx
@@ -4,10 +4,11 @@ import { AnimatePresence } from 'framer-motion'
 
 import { MotionBox } from '~components/motion'
 
-import { EditFieldDrawer } from './EditFieldDrawer/EditFieldDrawer'
-import { DrawerTabs, useBuilderDrawer } from './BuilderDrawerContext'
+import { DrawerTabs, useBuilderDrawer } from '../../BuilderDrawerContext'
+import { activeFieldSelector, useEditFieldStore } from '../../editFieldStore'
+
 import { CreateFieldDrawer } from './CreateFieldDrawer'
-import { activeFieldSelector, useEditFieldStore } from './editFieldStore'
+import { EditFieldDrawer } from './EditFieldDrawer'
 
 const DRAWER_MOTION_PROPS = {
   initial: { width: 0 },
