+++ frontend/src/features/admin-form-builder/components/BuilderDrawer/EditFieldDrawer/index.ts
@@ -1,0 +1,1 @@
+export { EditFieldDrawer } from './EditFieldDrawer'
