+++ frontend/src/features/admin-form-builder/components/BuilderSidebar/DrawerTabIcon.tsx
