+++ frontend/src/features/admin-form-builder/components/BuilderSidebar/index.ts
@@ -1,0 +1,1 @@
+export * from './BuilderSidebar'
