+++ frontend/src/features/admin-form-builder/components/FieldRow/index.ts
@@ -1,0 +1,1 @@
+export { FieldRowContainer as default } from './FieldRowContainer'
