+++ frontend/tsconfig.json
@@ -2,7 +2,11 @@
   "extends": "./tsconfig.paths.json",
   "compilerOptions": {
     "target": "es5",
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
     "allowJs": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
@@ -18,5 +22,8 @@
     "jsx": "react-jsx",
     "downlevelIteration": true
   },
-  "include": ["src/**/*", ".storybook/**/*"]
+  "include": [
+    "src/**/*",
+    ".storybook/**/*"
+  ]
 }
